OCR/AttentionOCR_flask/requirements.txt
numpy==1.17.4
Werkzeug==0.15.5
six==1.15.0
Flask_Bootstrap==3.3.7.1
tensorflow==1.13.1
Shapely==1.6.4.post2
opencv_python==4.4.0.46
tensorpack==0.9.8
Flask==1.1.1
matplotlib==3.0.3
scikit_image==0.16.2
Pillow==8.1.0
pycocotools==2.0.2
skimage==0.0
